+++ src/Layer_Test.h
@@ -22,6 +22,6 @@ public:
   ~TestLayer() = default;
   virtual void OnAttach() override;
   virtual void OnDetach() override;
-  virtual void OnUpdate(const double timeStamp) override;
+  virtual void OnUpdate(const float timeStamp) override;
   virtual void OnEvent(GLCore::Event &e) override;
 };
